+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComboBox.java
@@ -22,7 +22,6 @@ import is.codion.common.value.Value;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
-import is.codion.framework.i18n.FrameworkMessages;
 import is.codion.swing.common.model.component.combobox.FilteredComboBoxModel.ItemFinder;
 import is.codion.swing.common.ui.component.Components;
 import is.codion.swing.common.ui.component.combobox.ComboBoxBuilder;
@@ -210,28 +209,11 @@ public final class EntityComboBox extends JComboBox<Entity> {
 
     private DefaultBuilder(EntityComboBoxModel comboBoxModel, Value<Entity> linkedValue) {
       super(comboBoxModel, linkedValue);
-      popupMenuControl(comboBox -> Control.builder(new ForceRefreshCommand(comboBoxModel))
-              .name(FrameworkMessages.refresh())
-              .build());
     }
 
     @Override
     protected EntityComboBox createComboBox() {
       return new EntityComboBox((EntityComboBoxModel) comboBoxModel);
-    }
-  }
-
-  private static final class ForceRefreshCommand implements Control.Command {
-
-    private final EntityComboBoxModel comboBoxModel;
-
-    private ForceRefreshCommand(EntityComboBoxModel comboBoxModel) {
-      this.comboBoxModel = comboBoxModel;
-    }
-
-    @Override
-    public void perform() {
-      comboBoxModel.forceRefresh();
     }
   }
 }